+++ document_page/data/document_page.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<openerp>
+<odoo>
     <data noupdate="1">
         <record id="wiki_wiki_main" model="document.page">
             <field name="name">The OpenERP wiki</field>
@@ -8,8 +8,8 @@
             <field name="index">1</field>
             <field name="summary">Initial Page</field>
             <field  name="content">==The OpenERP wiki==
-            	
-[[File:http://www.openerp.com/sites/all/themes/openerp/logo.png OpenERP]]            	
+
+[[File:http://www.openerp.com/sites/all/themes/openerp/logo.png OpenERP]]
 
 The OpenERP wiki allows you to manage your enterprise's contents using wiki
 restructured texts. This module provides a collaborative way to manage internal
@@ -38,4 +38,4 @@ FAQs, quality manuals, technical references, etc.
             <field name="parent_id" ref="wiki_groups_wikiformatting0"/>
         </record>
     </data>
-</openerp>
+</odoo>
