+++ document_page/security/document_page_security.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
 <data noupdate="0">
     <record id="base.group_document_user" model="res.groups">
         <field name="users" eval="[(4, ref('base.user_root'))]"/>
     </record>
 </data>
-</openerp>
+</odoo>
