+++ document_page/wizard/document_page_create_menu.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
         <!-- Create Menu From view -->
         <record id="view_wiki_create_menu" model="ir.ui.view">
@@ -38,4 +38,4 @@
         </record>
 
     </data>
-</openerp>
+</odoo>
