+++ document_page/wizard/document_page_show_diff.py
@@ -1,3 +1,4 @@
+"""DocumentPageShowDiff model."""
 # -*- coding: utf-8 -*-
 ##############################################################################
 #
@@ -22,12 +23,13 @@ from openerp import models, fields, _
 from openerp import exceptions
 
 
-class showdiff(models.TransientModel):
-    """ Display Difference for History """
+class DocumentPageShowDiff(models.TransientModel):
+    """Display Difference for History."""
 
     _name = 'wizard.document.page.history.show_diff'
 
     def get_diff(self):
+        """Return the Difference between two document."""
         history = self.env["document.page.history"]
         ids = self.env.context.get('active_ids', [])
 
@@ -57,5 +59,3 @@ class showdiff(models.TransientModel):
         readonly=True,
         default=get_diff
     )
-
-# vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
