+++ document_page/wizard/document_page_show_diff.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <!-- Create Index Form view  -->
@@ -44,4 +44,4 @@
             view_type="form"/>
 
     </data>
-</openerp>
+</odoo>
